requirements.txt
python-chess>=1.999
pytest>=8.0
ruff>=0.3
